// src/lib/enums.ts
export enum QuestionType {
  MULTIPLE_CHOICE = 'multiple_choice',
  FILE_UPLOAD = 'file_upload',
  INPUT_BOX = 'text',
  CHECKBOXES = 'radio',
  DROPDOWN = 'single_choice',
  MULTI_INBOX = 'multi_inbox',
}

export enum TextInputType {
  TEXT = 'text',
  NUMBER = 'number',
  EMAIL = 'email',
  DATE = 'date',
  DATETIME = 'datetime',
  PHONE = 'phone',
  TAGS = 'tags',
}